Refactor propTypes
