+++ react/props/productKitItemProps.js
